Debugging grid 6x7 with adding more precision.

+++ voronoi/graph/algebra.py
@@ -26,7 +26,7 @@ class Algebra:
     def line_ray_intersection_point(ray_orig, ray_end, point_1, point_2):
         # Convert to numpy arrays
         orig = np.array(ray_orig, dtype=np.float)
-        end = np.array(ray_end)
+        end = np.array(ray_end, dtype=np.float)
         direction = np.array(Algebra.norm(end - orig), dtype=np.float)
         point_1 = np.array(point_1, dtype=np.float)
         point_2 = np.array(point_2, dtype=np.float)
